requirements.txt
# PokerGPT Dependencies
# Install with: pip install -r requirements.txt

google-genai>=1.0.0  # Google Gemini API client
python-dotenv>=1.0.0 # Load .env configuration files
streamlit>=1.30.0    # Web UI
